refactor: breakout code into `src/actions`
