+++ src/app.ts
@@ -1,19 +1,8 @@
 import CoinGeckoAPI from '@crypto-coffee/coingecko-api';
 import { saveCoinData } from './actions/saveCoinData';
-import { logError, logSuccess, format } from './utils';
-
-export type ExportData = Record<string, string | number>;
-
-interface Flags {
-  price: string[];
-  priceChange: boolean;
-  volume: boolean;
-  high: boolean;
-  low: boolean;
-  ath: boolean;
-  athChange: boolean;
-  save: string;
-}
+import { logError } from './utils';
+import { priceStats } from './actions/priceStats';
+import type { ExportData, Flags } from './constants';
 
 export const app = async (action: string, flags: Record<string, unknown>) => {
   const { price, priceChange, volume, high, low, ath, athChange, save } =
@@ -45,38 +34,6 @@ export const app = async (action: string, flags: Record<string, unknown>) => {
       ath: athPrice,
       ath_change_percentage: athPercent
     } of result) {
-      const priceRes = `${name}: ${format(current_price)}`;
-      let priceChangeRes;
-      let volumeRes;
-      let highRes;
-      let lowRes;
-      let athRes;
-      let athChangeRes;
-
-      if (priceChange) {
-        priceChangeRes = `change (24H): ${percent24h.toFixed(2)}%`;
-      }
-
-      if (high) {
-        highRes = `high (24H): ${format(high_24h)}`;
-      }
-
-      if (low) {
-        lowRes = `low (24H): ${format(low_24h)}`;
-      }
-
-      if (volume) {
-        volumeRes = `volume (24H): ${format(total_volume)}`;
-      }
-
-      if (ath) {
-        athRes = `ATH: ${format(athPrice)}`;
-      }
-
-      if (athChange) {
-        athChangeRes = `ATH (%): ${athPercent.toFixed(2)}%`;
-      }
-
       exportCoinData.push({
         name,
         current_price,
@@ -88,18 +45,18 @@ export const app = async (action: string, flags: Record<string, unknown>) => {
         ath_change_percentage: athPercent
       });
 
-      logSuccess(
-        [
-          priceRes,
-          priceChangeRes,
-          volumeRes,
-          highRes,
-          lowRes,
-          athRes,
-          athChangeRes
-        ]
-          .filter(Boolean)
-          .join(' - ')
+      priceStats(
+        {
+          name,
+          current_price,
+          total_volume,
+          high_24h,
+          low_24h,
+          percent24h,
+          athPrice,
+          athPercent
+        },
+        { price, priceChange, high, low, volume, ath, athChange }
       );
     }
